update tasks, refactor svelte files
