+++ ui/src/lib/features/views/index.ts
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2024-Present The UDS Authors
+
+export { default as Catalog } from './catalog/component.svelte';
