+++ ui/src/routes/apps/+page.svelte
@@ -1,19 +1,5 @@
 <script lang="ts">
-	import { base } from '$app/paths';
-	import type { PageData } from './$types';
-
-	export let data: PageData;
+	import { Catalog } from '$lib/features/views';
 </script>
 
-<div class="h-screen overflow-y-auto">
-	<div class="w-full max-w-full px-4 py-4 sm:py-8 md:py-12 lg:py-16">
-		<div class="mx-auto max-w-3xl">
-			<div class="mx-auto mb-8 max-w-sm drop-shadow-2xl sm:mb-12 md:mb-16">
-				<img src="{base}/images/dougandserver.svg" alt="UDS Logo" class="h-auto w-full" />
-			</div>
-			<h1 class="mb-4 text-center text-2xl font-bold sm:text-3xl md:text-4xl lg:text-5xl">
-				UDS applications ({data.apps.length}) coming soon!
-			</h1>
-		</div>
-	</div>
-</div>
+<Catalog />
