+++ ui/vite.config.ts
@@ -6,15 +6,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig(({ mode }) => ({
 	plugins: [sveltekit()],
-	server: {
-		proxy: {
-			// Proxy all requests starting with /api to the go server
-			'/api': {
-				target: 'http://localhost:8080',
-				changeOrigin: true
-			}
-		}
-	},
 	test: {
 		include: ['src/**/*.test.{js,ts}'],
 		environment: 'jsdom',
